requirements.txt
djitellopy==2.5.0
vosk==0.3.45
mediapipe==0.10.21
opencv-python==4.8.1.78
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
sounddevice>=0.4.4
keyboard>=1.13.0
python-dotenv>=0.19.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.5.0 